+++ apps/site-app/src/containers/Sidebar/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { getAvatarLink } from '@thecointech/shared/components/Avatars';
 import { AccountMap } from '@thecointech/redux-accounts';
-import { SidebarItemsReducer } from '@thecointech/shared/containers/PageSidebar/reducer';
-import type { SidebarLink } from '@thecointech/shared/containers/PageSidebar/types';
+import { SidebarItemsReducer, type SidebarLink } from '@thecointech/shared/containers/PageSidebar';
 import { defineMessages, useIntl } from 'react-intl';
 import { isLocal } from '@thecointech/signers';
 
-import { getOpenRoutes, type AuthPathKey } from '../App/Routes';
+import type { AuthPathKey } from '../App/Routes';
+import { getOpenRoutes } from '../App/Routes.open';
 import { useLocation } from 'react-router';
 
 const translations = defineMessages({